+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Calendar.js - Change Log:
 
 ### Version 0.9.7:
+- Added repeat Every 2 Weeks support (please update your events, the IDs for the others have been changed.  Sorry).
 - 
 
 ### Version 0.9.6: